Clang build fix 12 Final test  fix

+++ src/Protocol/Protocol_1_13.cpp
@@ -89,7 +89,7 @@ void cProtocol_1_13::SendMapData(const cMap & a_Map, int a_DataStartX, int a_Dat
 		Pkt.WriteVarInt32(static_cast<UInt32>(a_Map.GetDecorators().size()));
 		for (const auto & Decorator : a_Map.GetDecorators())
 		{
-			Pkt.WriteVarInt32(static_cast<Int32>(Decorator.GetType()));
+			Pkt.WriteVarInt32(static_cast<UInt32>(Decorator.GetType()));
 			Pkt.WriteBEUInt8(static_cast<UInt8>(Decorator.GetPixelX()));
 			Pkt.WriteBEUInt8(static_cast<UInt8>(Decorator.GetPixelZ()));
 			Pkt.WriteBEUInt8(static_cast<UInt8>(Decorator.GetRot()));
@@ -282,7 +282,7 @@ void cProtocol_1_13::SendSpawnEntity(const cEntity & a_Entity)
 	{
 		const auto & Block = static_cast<const cFallingBlock &>(a_Entity);
 		auto NumericBlock = GetProtocolBlockType(Block.GetBlock());
-		EntityData = NumericBlock;
+		EntityData = static_cast<Int32>(NumericBlock);
 	}
 	else if (a_Entity.IsFloater())
 	{
